backend/requirements.txt
# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0

# MCP Integration
mcp>=0.1.0

# LLM Integration
openai>=1.0.0
anthropic>=0.18.0

# Environment
python-dotenv>=1.0.0

# Geospatial
geopandas>=0.14.0
shapely>=2.0.0

# Testing
pytest>=7.4.0
pytest-flask>=1.3.0

# Development
black>=23.0.0
flake8>=6.0.0
